move `Server` module to `/lib` directory
